Move to a local build system for QSimpleUpdater
As we did with LUA, QSimpleUpdater has a local CMake build system, which
we just include. No port is needed.

I am abandoning the idea of upstreaming the CMake support - as I see for
example that LUA is not intersted at this at all. I think SCintilla are
in a similar position.

If anyone wants to port these libraries to use CMake, they can adpot
this code from ours.