Simplify build system
Now that editor-config-qt cmake support is upstream we can use the
upstream repo, instead of the previous port.

I also removed all other forks and instead made a local cmake build
system (local to the project, a simple include file). Now we are left
with a single fork - which is necesary (upstream uses mervurial, and
demands python for generting some files on build time - my fork has that
as part of the tree).

All projects point to the newewst code available on their git - code
seems to compile and run, so... yap, ship it.

Fixed the app name and also remove lots of craft from the main
CMakeFile.

Fixed the workflow to use the *.so files generated at build, and
disabled the OSX build for now.